move registering reducers
