+++ src/register/reducer.js
@@ -3,13 +3,13 @@
 import { combineReducers } from 'redux';
 import type { Reducer } from 'redux';
 
-import createReducers from '.';
+import createReducers from '../reducers/index';
 
 type Reducers = { [key: string]: Reducer<ThymeStore, *> };
 
 const registeredReducers: { [path: string]: Reducers[] } = {};
 
-export function registerReducer(store: ThymeStore, path: string, reducers: Reducers) {
+export function register(store: ThymeStore, path: string, reducers: Reducers) {
   if (!registeredReducers[path]) {
     registeredReducers[path] = [];
   }
@@ -22,10 +22,7 @@ export function registerReducer(store: ThymeStore, path: string, reducers: Reduc
   store.replaceReducer(createReducers());
 }
 
-export function createExtendableReducer(
-  path: string,
-  reducers: Reducers,
-) {
+export function create(path: string, reducers: Reducers) {
   const extraReducers = registeredReducers[path] || [];
 
   const newReducers = {
