+++ src/sections/Settings/reducers/index.js
@@ -1,11 +1,11 @@
 // @flow
 
-import { createExtendableReducer } from 'reducers/registerReducer';
+import { create } from 'register/reducer';
 
 import createRoundingReducer from './rounding';
 import createTimesheetReducer from './timesheet';
 
-export default () => createExtendableReducer('settings', {
+export default () => create('settings', {
   rounding: createRoundingReducer(),
   timesheet: createTimesheetReducer(),
 });
