+++ src/sections/TimeSheet/reducers/time.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { createExtendableReducer } from 'reducers/registerReducer';
+import { create } from 'register/reducer';
 
 function id(state = null, action) {
   switch (action.type) {
@@ -84,7 +84,7 @@ function updatedAt(state = new Date(), action) {
   }
 }
 
-export default () => createExtendableReducer('timesheet.time', {
+export default () => create('timesheet.time', {
   id,
   project,
   start,
